# frontend/vue-project/merged-api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://10.239.105.110:11082",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v2/storage/Download": {
      "get": {
        "tags": [
          "upload-v-2-service"
        ],
        "operationId": "generateDownloadUrl",
        "parameters": [
          {
            "name": "virtualFileId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/shares/verify/{shareCode}": {
      "post": {
        "tags": [
          "分享模块"
        ],
        "summary": "访问分享链接",
        "description": "使用提取码访问分享链接",
        "operationId": "verifySharedFile",
        "parameters": [
          {
            "name": "shareCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/share": {
      "post": {
        "tags": [
          "分享模块"
        ],
        "summary": "创建分享链接",
        "description": "根据请求体创建分享链接",
        "operationId": "shareFile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultCreateShareResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/save": {
      "post": {
        "tags": [
          "分享模块"
        ],
        "summary": "转存文件",
        "description": "访问者将文件转存到自己的网盘中，需要指定目标文件夹",
        "operationId": "saveSharedFile",
        "parameters": [
          {
            "name": "Share-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/list": {
      "get": {
        "tags": [
          "分享模块"
        ],
        "summary": "查询分享文件列表",
        "description": "查询属于自己的分享链接列表",
        "operationId": "listSharedFile",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "description": "页码",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1001
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "页大小",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 50
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageResultShareLinkVO"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/info": {
      "get": {
        "tags": [
          "分享模块"
        ],
        "summary": "访问他人分享详情",
        "description": "访问他人分享链接具体内容",
        "operationId": "getShareInfo",
        "parameters": [
          {
            "name": "Share-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parentId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "rootId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListVirtualFileVO"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/download": {
      "get": {
        "tags": [
          "分享模块"
        ],
        "summary": "下载文件",
        "description": "访问者直接在网页中下载文件，返回一个下载链接",
        "operationId": "downloadSharedFile",
        "parameters": [
          {
            "name": "Share-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "rootId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/cancel": {
      "delete": {
        "tags": [
          "分享模块"
        ],
        "summary": "删除分享",
        "description": "删除指定的分享链接",
        "operationId": "cancelSharedFile",
        "parameters": [
          {
            "name": "shareId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Result": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应状态码，1表示成功",
            "format": "int32",
            "example": 1
          },
          "msg": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          }
        },
        "description": "通用响应对象"
      },
...
      "FileItemType": {
        "type": "string",
        "description": "文件/文件夹类型",
        "example": "FILE",
        "enum": [
          "FILE",
          "FOLDER"
        ]
      },
      "ItemIdentity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "文件/文件夹ID",
            "format": "int64",
            "example": 1001
          },
          "type": {
            "$ref": "#/components/schemas/FileItemType"
          }
        },
        "description": "简易对象实体",
        "example": [
          {
            "id": 100,
            "type": "FILE"
          },
          {
            "id": 200,
            "type": "FOLDER"
          }
        ]
      },
...
      "ResultString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应状态码，1表示成功",
            "format": "int32",
            "example": 1
          },
          "msg": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "type": "string",
            "description": "响应数据"
          }
        },
        "description": "通用响应对象"
      },
...
      "ResultListVirtualFileVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应状态码，1表示成功",
            "format": "int32",
            "example": 1
          },
          "msg": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "type": "array",
            "description": "响应数据",
            "items": {
              "$ref": "#/components/schemas/VirtualFileVO"
            }
          }
        },
        "description": "通用响应对象"
      },
      "VirtualFileVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "文件/文件夹ID",
            "format": "int64",
            "example": 1001
          },
          "name": {
            "type": "string",
            "description": "文件/文件夹名称",
            "example": "name"
          },
          "size": {
            "type": "integer",
            "description": "文件大小（字节）",
            "format": "int64",
            "example": 1024
          },
          "updateTime": {
            "type": "string",
            "description": "最后更新时间",
            "format": "date-time"
          },
          "type": {
            "$ref": "#/components/schemas/FileItemType"
          },
          "parentId": {
            "type": "integer",
            "description": "父目录ID",
            "format": "int64",
            "example": 1001
          }
        },
        "description": "用于前端展示的视图对象"
      },
...
      "CreateShareRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "提取码",
            "example": "abcde"
          },
          "expireTime": {
            "type": "string",
            "description": "过期时间",
            "format": "date-time"
          },
          "displayName": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "description": "分享的文件/文件夹对象",
            "example": [
              {
                "id": 100,
                "type": "FILE"
              },
              {
                "id": 200,
                "type": "FOLDER"
              }
            ],
            "items": {
              "$ref": "#/components/schemas/ShareItem"
            }
          }
        },
        "description": "创建分享的请求体"
      },
      "ShareItem": {
        "type": "object",
        "properties": {
          "targetId": {
            "type": "integer",
            "description": "文件/文件夹ID",
            "format": "int64",
            "example": 1001
          },
          "targetType": {
            "$ref": "#/components/schemas/FileItemType"
          }
        },
        "description": "分享的文件/文件夹对象",
        "example": [
          {
            "id": 100,
            "type": "FILE"
          },
          {
            "id": 200,
            "type": "FOLDER"
          }
        ]
      },
      "CreateShareResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "分享链接ID",
            "format": "int64",
            "example": 1001
          },
          "shareCode": {
            "type": "string",
            "description": "分享链接唯一码(前端拼接URL)",
            "example": "aaabbbccc"
          },
          "password": {
            "type": "string",
            "description": "提取码",
            "example": "abcde"
          },
          "expireTime": {
            "type": "string",
            "description": "过期时间",
            "format": "date-time"
          },
          "createTime": {
            "type": "string",
            "description": "创建分享时间",
            "format": "date-time"
          }
        },
        "description": "创建分享的响应体 "
      },
      "ResultCreateShareResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应状态码，1表示成功",
            "format": "int32",
            "example": 1
          },
          "msg": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/CreateShareResponse"
          }
        },
        "description": "通用响应对象"
      },
      "TransferRequest": {
        "type": "object",
        "properties": {
          "targetFolderId": {
            "type": "integer",
            "description": "目标文件夹id（访问者视角下的文件列表）",
            "format": "int64",
            "example": 1001
          },
          "items": {
            "type": "array",
            "description": "待转存的文件/文件夹标识",
            "example": [
              {
                "id": 100,
                "type": "FILE"
              },
              {
                "id": 200,
                "type": "FOLDER"
              }
            ],
            "items": {
              "$ref": "#/components/schemas/ItemIdentity"
            }
          },
          "rootId": {
            "type": "integer",
            "description": "源文件夹的顶级文件夹id（访问者视角下的文件列表）",
            "format": "int64",
            "example": 1001
          }
        }
      },
      "PageResultShareLinkVO": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShareLinkVO"
            }
          }
        },
        "description": "响应数据"
      },
      "ResultPageResultShareLinkVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应状态码，1表示成功",
            "format": "int32",
            "example": 1
          },
          "msg": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/PageResultShareLinkVO"
          }
        },
        "description": "通用响应对象"
      },
      "ShareLinkVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "分享链接ID",
            "format": "int64",
            "example": 1001
          },
          "shareCode": {
            "type": "string",
            "description": "分享链接唯一码(前端拼接URL)",
            "example": "aaabbbccc"
          },
          "password": {
            "type": "string",
            "description": "提取码",
            "example": "abcde"
          },
          "expireTime": {
            "type": "string",
            "description": "过期时间",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/ShareStatus"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "最后更新时间",
            "format": "date-time"
          },
          "displayName": {
            "type": "string",
            "description": "分享链接显示名称",
            "example": "分享"
          }
        },
        "description": "用于前端返回的视图对象"
      },
      "ShareStatus": {
        "type": "string",
        "description": "分享链接状态",
        "example": "ACTIVE",
        "enum": [
          "NO_ACTIVE",
          "ACTIVE"
        ]
      },
...
    }
  }
}